aider: Fix width of conversation view.
# Aider chat conversation:

USER: fix width of conversation view

ASSISTANT: To fix the width of the conversation view, we need to adjust the width calculation in the `RenderChat` function in `view.go`.

Here is the *SEARCH/REPLACE* block:

view.go
```go
<<<<<<< SEARCH
		Width(terminalWidth - (4 * borderWidth)).
=======
		Width(terminalWidth - (2 * borderWidth)).
>>>>>>> REPLACE
```

This change will ensure that the width of the conversation view is calculated correctly.